fix(lint): Move normalize-latex to infra layer to fix import restriction
Infra layer cannot import from UI layer. Moved normalizeLatexDelimiters
to src/infra/utils/normalize-latex.ts and re-exported from original
location to avoid breaking existing UI imports.
